// frontend/app/openapi-generator/models/BulkRequestBody.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Enduro API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Request body for bulk.
 * @export
 * @interface BulkRequestBody
 */
export interface BulkRequestBody {
    /**
     * 
     * @type {string}
     * @memberof BulkRequestBody
     */
    operation: BulkRequestBodyOperationEnum;
    /**
     * 
     * @type {number}
     * @memberof BulkRequestBody
     */
    size?: number;
    /**
     * 
     * @type {string}
     * @memberof BulkRequestBody
     */
    status: BulkRequestBodyStatusEnum;
}

/**
* @export
* @enum {string}
*/
export enum BulkRequestBodyOperationEnum {
    Retry = 'retry',
    Cancel = 'cancel',
    Abandon = 'abandon',
    UnknownDefaultOpenApi = '11184809'
}
/**
* @export
* @enum {string}
*/
export enum BulkRequestBodyStatusEnum {
    New = 'new',
    InProgress = 'in progress',
    Done = 'done',
    Error = 'error',
    Unknown = 'unknown',
    Queued = 'queued',
    Pending = 'pending',
    Abandoned = 'abandoned',
    UnknownDefaultOpenApi = '11184809'
}


/**
 * Check if a given object implements the BulkRequestBody interface.
 */
export function instanceOfBulkRequestBody(value: object): value is BulkRequestBody {
    if (!('operation' in value) || value['operation'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    return true;
}

export function BulkRequestBodyFromJSON(json: any): BulkRequestBody {
    return BulkRequestBodyFromJSONTyped(json, false);
}

export function BulkRequestBodyFromJSONTyped(json: any, ignoreDiscriminator: boolean): BulkRequestBody {
    if (json == null) {
        return json;
    }
    return {
        
        'operation': json['operation'],
        'size': json['size'] == null ? undefined : json['size'],
        'status': json['status'],
    };
}

export function BulkRequestBodyToJSON(json: any): BulkRequestBody {
    return BulkRequestBodyToJSONTyped(json, false);
}

export function BulkRequestBodyToJSONTyped(value?: BulkRequestBody | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'operation': value['operation'],
        'size': value['size'],
        'status': value['status'],
    };
}