// frontend/app/openapi-generator/models/BulkStatusResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Enduro API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BulkStatusResult
 */
export interface BulkStatusResult {
    /**
     * 
     * @type {Date}
     * @memberof BulkStatusResult
     */
    closedAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof BulkStatusResult
     */
    runId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof BulkStatusResult
     */
    running: boolean;
    /**
     * 
     * @type {Date}
     * @memberof BulkStatusResult
     */
    startedAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof BulkStatusResult
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof BulkStatusResult
     */
    workflowId?: string;
}

/**
 * Check if a given object implements the BulkStatusResult interface.
 */
export function instanceOfBulkStatusResult(value: object): value is BulkStatusResult {
    if (!('running' in value) || value['running'] === undefined) return false;
    return true;
}

export function BulkStatusResultFromJSON(json: any): BulkStatusResult {
    return BulkStatusResultFromJSONTyped(json, false);
}

export function BulkStatusResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): BulkStatusResult {
    if (json == null) {
        return json;
    }
    return {
        
        'closedAt': json['closed_at'] == null ? undefined : (new Date(json['closed_at'])),
        'runId': json['run_id'] == null ? undefined : json['run_id'],
        'running': json['running'],
        'startedAt': json['started_at'] == null ? undefined : (new Date(json['started_at'])),
        'status': json['status'] == null ? undefined : json['status'],
        'workflowId': json['workflow_id'] == null ? undefined : json['workflow_id'],
    };
}

export function BulkStatusResultToJSON(json: any): BulkStatusResult {
    return BulkStatusResultToJSONTyped(json, false);
}

export function BulkStatusResultToJSONTyped(value?: BulkStatusResult | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'closed_at': value['closedAt'] == null ? value['closedAt'] : value['closedAt'].toISOString(),
        'run_id': value['runId'],
        'running': value['running'],
        'started_at': value['startedAt'] == null ? value['startedAt'] : value['startedAt'].toISOString(),
        'status': value['status'],
        'workflow_id': value['workflowId'],
    };
}